add a e1m1 only truncation of id1_er
Does not include `start`.  Calling `map` directly on a level is slightly
different to entering the level via start (or presumably any other map),
as the client spawns one frame earlier in the `map` case.  Since we want
the truncated version (which uses map) to be transposable into the full
version, then we must delay the start by a frame.  I've added a
tas_connect_delay cvar to this end.

In addition, we need to set up the RNG seed.  The best place to do this
seems to be in SV_SpawnServer.  I've added a print statement here for
grabbing the seed from id1_er, and a new command for setting the seed in
the truncated .qtas file. (The old command postpones setting the seed
until the next frame which doesn't work since the RNG is used inside
SV_SpawnServer itself.)

This solution only really works for the first level of each episode,
since all the other levels have carried over state.  A more general
solution would probably be to get save_state / load_state working.